[batch] refactor expected value calculations
