+++ src/batch/monitor.tsx
@@ -4,6 +4,7 @@ import { ALLOC_ID, ALLOC_ID_ARG, MEM_TAG_FLAGS } from "services/client/memory_ta
 import { MONITOR_PORT, Lifecycle, Message as MonitorMessage } from "batch/client/monitor";
 
 import { expectedValuePerRamSecond } from "batch/expected_value";
+import { CONFIG } from "batch/config";
 
 import { DiscoveryClient } from "services/client/discover";
 import { TaskSelectorClient } from "batch/client/task_selector";
@@ -341,7 +342,7 @@ export function hostInfo(ns: NS, target: string, targetThreads: TargetThreads): 
     const secPlus = sec - minSec;
 
     const harvestMoney = targetThreads.harvestMoney;
-    const expectedValue = expectedValuePerRamSecond(ns, target);
+    const expectedValue = expectedValuePerRamSecond(ns, target, CONFIG.maxHackPercent);
 
     return {
         name: target,
