+++ src/batch/task_selector.ts
@@ -326,7 +326,7 @@ class TaskSelector {
             .filter(h => canHarvest(this.ns, h) && worthHarvesting(this.ns, h))
             .map(h => ({
                 host: h,
-                value: expectedValuePerRamSecond(this.ns, h),
+                value: expectedValuePerRamSecond(this.ns, h, CONFIG.maxHackPercent),
                 ...calculateBatchLogistics(this.ns, h)
             }))
             .sort((a, b) => b.value - a.value);
@@ -539,5 +539,5 @@ function canHarvest(ns: NS, hostname: string) {
 }
 
 function worthHarvesting(ns: NS, hostname: string) {
-    return expectedValuePerRamSecond(ns, hostname) > CONFIG.expectedValueThreshold;
+    return expectedValuePerRamSecond(ns, hostname, CONFIG.maxHackPercent) > CONFIG.expectedValueThreshold;
 }
